+++ src/fr/maxlego08/menu/ZInventoryManager.java
@@ -40,6 +40,7 @@ import fr.maxlego08.menu.loader.actions.MessageLoader;
 import fr.maxlego08.menu.loader.actions.PlayerCommandLoader;
 import fr.maxlego08.menu.loader.actions.ShopkeeperLoader;
 import fr.maxlego08.menu.loader.actions.SoundLoader;
+import fr.maxlego08.menu.loader.actions.TitleLoader;
 import fr.maxlego08.menu.loader.permissible.ItemPermissibleLoader;
 import fr.maxlego08.menu.loader.permissible.PermissionPermissibleLoader;
 import fr.maxlego08.menu.loader.permissible.PlaceholderPermissibleLoader;
@@ -269,6 +270,7 @@ public class ZInventoryManager extends ZUtils implements InventoryManager {
         buttonManager.registerAction(new ConsoleCommandLoader());
         buttonManager.registerAction(new fr.maxlego08.menu.loader.actions.BackLoader(this.plugin));
         buttonManager.registerAction(new ShopkeeperLoader());
+        buttonManager.registerAction(new TitleLoader());
 
         // Loading ButtonLoader
         // The first step will be to load the buttons in the plugin, so each
@@ -277,7 +279,7 @@ public class ZInventoryManager extends ZUtils implements InventoryManager {
         buttonManager.register(new NoneLoader(this.plugin, ZNoneButton.class, "none"));
         buttonManager.register(new NoneLoader(this.plugin, ZNoneButton.class, "none_slot"));
         buttonManager.register(new NoneLoader(this.plugin, ZNoneButton.class, "perform_command"));
-        buttonManager.register(new fr.maxlego08.menu.button.loader.InventoryLoader(this.plugin, this));
+        buttonManager.register(new fr.maxlego08.menu.button.loader.InventoryLoader(this.plugin, this, this.plugin.getCommandManager()));
         buttonManager.register(new BackLoader(this.plugin, this));
         buttonManager.register(new HomeLoader(this.plugin, this));
         buttonManager.register(new NextLoader(this.plugin, this));