+++ src/fr/maxlego08/menu/api/pattern/Pattern.java
@@ -31,4 +31,6 @@ public interface Pattern {
      */
     Collection<Button> getButtons();
 
+    boolean enableMultiPage();
+
 }