+++ src/fr/maxlego08/menu/loader/actions/BroadcastSoundLoader.java
@@ -5,26 +5,14 @@ import fr.maxlego08.menu.api.loader.ActionLoader;
 import fr.maxlego08.menu.api.requirement.Action;
 import fr.maxlego08.menu.api.utils.TypedMapAccessor;
 import fr.maxlego08.menu.requirement.actions.BroadcastSoundAction;
-import fr.maxlego08.menu.requirement.actions.SoundAction;
 import fr.maxlego08.menu.sound.ZSoundOption;
 
 import java.io.File;
-import java.util.Map;
 
-public class BroadcastSoundLoader implements ActionLoader {
+public class BroadcastSoundLoader extends SoundLoader {
 
     @Override
     public String getKey() {
         return "broadcast_sound";
-    }
-
-    @Override
-    public Action load(String path, TypedMapAccessor accessor, File file) {
-
-        String sound = accessor.getString("sound");
-        float pitch = accessor.getFloat("pitch", 1f);
-        float volume = accessor.getFloat("volume", 1f);
-        XSound xSound = sound == null || sound.isEmpty() ? null : XSound.matchXSound(sound).orElse(null);
-        return new BroadcastSoundAction(new ZSoundOption(xSound, pitch, volume));
     }
 }