+++ src/fr/maxlego08/menu/pattern/ZPattern.java
@@ -10,11 +10,13 @@ public class ZPattern implements Pattern {
     private final String name;
     private final List<Button> buttons;
     private final int inventorySize;
+    private final boolean enableMultiPage;
 
-    public ZPattern(String name, List<Button> buttons, int inventorySize) {
+    public ZPattern(String name, List<Button> buttons, int inventorySize, boolean enableMultiPage) {
         this.name = name;
         this.buttons = buttons;
         this.inventorySize = inventorySize;
+        this.enableMultiPage = enableMultiPage;
     }
 
     @Override
@@ -30,5 +32,10 @@ public class ZPattern implements Pattern {
     @Override
     public List<Button> getButtons() {
         return buttons;
+    }
+
+    @Override
+    public boolean enableMultiPage() {
+        return this.enableMultiPage;
     }
 }