+++ changelog.md
@@ -24,14 +24,26 @@
 - [ ] Add the Pagination button, it will allow to take an input number, an output number and browse the value number
 - [ ] Add a way to display the list of online players in a button list
 - [ ] Add Citizen and [ZNPCsPlus](https://www.spigotmc.org/resources/znpcsplus.109380/) support to open an inventory by clicking on an NPC
-- [ ] Add a tool to transform ItemStack to base64 and vice versa. This tool can be used to all my plugins.
+- [x] Add a tool to transform ItemStack to base64 and vice versa. This tool can be used to all my plugins.
 - [ ] Fix the bug about the `clearInventory: true`. It does not restore inventory when the menu is closed.
 - [ ] Add the support for the `http://textures.minecraft.net/texture/e34969c2684e4f62d5f87875460441a9f849d296c01e4c621636bb6acda696f7` in the URL of a custom head.
 - [ ] Update pom.xml for add {projet.version} in plugin.yml
 - [ ] Add matrix support for slot (like this: https://abstractmenus.github.io/docs/general/item_format.html#way-4-matrix)
 - [ ] Adding more logs on the errors that can occur with custom items like ItemAdder, this will cause an error but the user will not have the information of why, for example when the item does not exist.
+- [ ] Create a new class for loading buttons to add more elements, like a boolean to check if the button needs an itemstack
 
 # Unreleased
+
+# 1.0.2.5
+
+- Add new lore as string support
+- Fix itemstack amount of 0
+- Support file with space
+- Improved API, added many new methods to make API more permissive.
+- Patterns can now be on multiple pages
+- Add custom sound to SoundOption
+- Added ability to add arguments and select page for INVENTORY button [#45](https://github.com/Maxlego08/zMenu/issues/45)
+- Change update interval to milliseconds [#37](https://github.com/Maxlego08/zMenu/issues/37)
 
 # 1.0.2.4